rename: biomon to combat
