--- src/components/charsheet/biomon/index.ts
@@ -1,1 +1,0 @@
-export { Biomonitor } from './Biomonitor';
