+++ src/components/charsheet/combat/BodyInfo.tsx
