+++ src/components/charsheet/combat/CharacterArmorPanel.tsx
