+++ src/components/charsheet/combat/CombatPanel.tsx
