+++ src/components/charsheet/combat/StabilizedControl.tsx
