+++ src/components/charsheet/combat/help/WoundHelpContent.tsx
