+++ src/components/charsheet/combat/index.ts
@@ -1,0 +1,1 @@
+export { CombatView } from './CombatView';
