+++ src/layouts/layout.astro
@@ -54,7 +54,7 @@ const links = {
           </div>
         </ul>
         <div class="nav-operational">
-          <div class="flex-between-baseline gap-4"><span class="status connected">Online</span><span class="text-desc">Biomonitor</span></div>
+          <div class="flex-between-baseline gap-4"><span class="status connected">Online</span><span class="text-desc">Combat</span></div>
           <div class="flex-between-baseline gap-4"><span class="status reconnect">Reconnection</span><span class="text-desc">Equipment</span></div>
           <a class="nav-link" href={links.charsheet} onclick="if(location.pathname.includes('charsheet')){event.preventDefault()}">Operational</a>
         </div>
