+++ src/scripts/combat/stats.ts
