+++ src/scripts/combat/types.ts
