+++ src/scripts/combat/wounds.ts
